lib: fix tests
In order for the tests to correctly run, a change was needed on the
`new` function prototype.

Instead of expecting an `std::env::Args` object we now expect a generic
type that implements an `Iterator` trait whose `Item` type is a `String`
struct.

This way we can use `vec![].into_iter()` on the tests to simulate the
command line arguments.