feature: add stdin as input
This commit adds the ability to read input also from
the `stdin` as well as a file.

The code that handles the matching was also moved into
a new function, which returns a boxed `Read` trait
enveloped in `Result`, in case it succeeds. In case of
failure, it will emit errors described by the document
of said function.